app/requirements.txt
bleach==4.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
docutils==0.14
Flask==2.0.2
Flask-Caching==1.10.1
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.1
isso==0.12.5
itsdangerous==2.1.0
Jinja2==3.0.3
lockfile==0.11.0
MarkupSafe==2.1.0
misaka==2.1.1
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pycparser==2.21
pyparsing==3.0.7
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
simplejson==3.2.0
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.0.3
xmltodict==0.13.0
zipp==3.7.0
cachetools==3.1.1
flask-cors
python-dotenv
mongoengine
uuid
mjml==0.8.0
jinja2
flask-admin==1.6.0
flask-mongoengine==1.0.0